Fix resource packs sometimes throwing error, inline audience
