+++ paper-server/patches/sources/net/minecraft/server/network/ServerCommonPacketListenerImpl.java.patch
@@ -63,7 +63,7 @@
 +        if (callback != null) {
 +            net.kyori.adventure.audience.Audience audience = switch (this) {
 +                case ServerGamePacketListenerImpl serverGamePacketListener -> serverGamePacketListener.getCraftPlayer();
-+                case ServerConfigurationPacketListenerImpl configurationPacketListener -> configurationPacketListener.paperConnection.getConfiguringPlayer();
++                case ServerConfigurationPacketListenerImpl configurationPacketListener -> configurationPacketListener.paperConnection.getAudience();
 +                default -> throw new IllegalStateException("Unexpected value: " + this);
 +            };
 +
