+++ paper-server/patches/sources/net/minecraft/server/network/ServerConfigurationPacketListenerImpl.java.patch
@@ -51,7 +51,7 @@
          this.configurationTasks.add(new JoinWorldTask());
          this.startNextTask();
      }
-@@ -105,6 +_,7 @@
+@@ -105,12 +_,13 @@
      @Override
      public void handleClientInformation(ServerboundClientInformationPacket packet) {
          this.clientInformation = packet.information();
@@ -59,6 +59,13 @@
      }
  
      @Override
+     public void handleResourcePackResponse(ServerboundResourcePackPacket packet) {
+         super.handleResourcePackResponse(packet);
+-        if (packet.action().isTerminal()) {
++        if (packet.action().isTerminal() && packet.id().equals(this.server.getServerResourcePack().map(MinecraftServer.ServerResourcePackInfo::id).orElse(null))) { // Paper - Ignore resource pack requests that are not vanilla
+             this.finishCurrentTask(ServerResourcePackConfigurationTask.TYPE);
+         }
+     }
 @@ -139,16 +_,30 @@
                  return;
              }
